add shipStore and authentication
