+++ package.json
@@ -15,9 +15,11 @@
 		"bootstrap-vue": "2.0.0-rc.11",
 		"firebase": "^5.2.0",
 		"ionicons": "^4.2.4",
+		"js-cookie": "^2.2.0",
+		"persist": "^0.2.7",
 		"vue": "2.5.16",
 		"vue-router": "3.0.1",
-		"vuex": "3.0.1",
+		"vuex": "^3.0.1",
 		"vuex-persistedstate": "2.4.2"
 	},
 	"devDependencies": {
