+++ src/vuex/modules/shipStore.js
@@ -1,20 +1,7 @@
-import firebase from 'firebase';
-
-const firebaseConfig = {
-	apiKey: 'AIzaSyAbm7-uLDKGBdBbYrrWzDutc2tNwdj3OTc',
-	authDomain: 'newagent-1cb47.firebaseapp.com',
-	databaseURL: 'https://newagent-1cb47.firebaseio.com',
-	projectId: 'newagent-1cb47',
-	storageBucket: 'newagent-1cb47.appspot.com',
-	messagingSenderId: '818684151536'
-};
-const app = firebase.initializeApp(firebaseConfig);
-const db = app.database();
-const matchesRef = db.ref('matches');
-
-const store = {
+export default {
+	namespaced: true,
 	state: {
-		matchesRef,
+		path: 'ship',
 		matchesCount: null,
 		messagesCount: null,
 		matchesUnread: null,
@@ -59,5 +46,3 @@ const store = {
 		messagesUnread: state => state.messagesUnread
 	}
 };
-
-export default store;
